+++ webui/react/src/services/api-ts-sdk/.swagger-codegen/VERSION
@@ -1,1 +1,1 @@
-2.4.14
+2.4.27
